Numbers (Factors): * Debug IntOperations exponent function * Update IntOperationsTest - move to factors package, add test cases to increase coverage
